command: make command more object-like
Update command to a more object like state by: creating a constructor,
giving it methods, and implement them by taking relevant code from
eeprom_utility.

Update parser to work with the constructor. This involved mostly a change
to the various parse_*() functions to return a value instead of inputting
straight into command, and slight review of the main parsing flow.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>