layout: reimplement update functions
This rewrite most notably eliminates enum layout_res, and ceases use
of struct offset_value_pair for by-byte changes. Now, by-byte changes
should be passed in the same strings_pair as by-field changes, where
the key is the offset and value is new byte value, both in string form.
This is in preparation for the elimination of field_data/byte_data
distinction.

Since the parser does not currently support this new format, by-byte
updates are disabled in command.c.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>