eeprom_util: move main() to the parser code
eeprom_utility.c now only runs the parser and executes the result.
It doesn't serve any purpose to keep this logic in its own file when
the code that creates the command can simply execute it on the spot
instead of returning it. Rename parse() to main() and just do the
execution together with the command creation.

An added benefit of this move is that new data allocation/deallocation
is more orderly: the space reserved for new data is deallocated in the
same place where it is allocated, as opposed to the previous situation
where somebody else had to clean up after the parser. This simplifies
free_command(), which now only cleans up after new_command().

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>