common: add base param to safe_strtoui()
Add a base param to safe_strtoui() so that it will be possible to
pick what kind of numerical input is acceptable. This is necessary
because byte updates can be described in any base, but most of the
byte fields are displayed as hexadecimal values, and so their update
data should also be interpreted as hexadecimal values. i.e.: mac
addresses are displayed as a0:b0:c0:d0:e0:f0, so in order to
update it you should type 0a:0b:0c:0d:0e:0f, not
0x0a:0x0b:0x0c:0x0d:0x0e:0x0f.

Signed-off-by: Nikita Kiryanov <nikita@compulab.co.il>