field: add a new __update_bin() function
The rational for the new function is that the old __update_bin_delim()
function heavily relies on the delimiter and does not handle the case
of an empty delimiter. In addition the "reversed" functionality of the
old __update_bin_delim() function is not implemented correctly and thus
stores the values in the wrong order thus violating the CompuLab EEPROM
layout specification.
The new function is used to properly update the binary and binary
reversed fields which do not have any delimiter between the byte values.
Also switch to using the new function with the serial number field.

This also changes the user API to conform with commit
63efc8b (field: remove space separation)
and provide the same syntax for updating the serial number field as it
is used to display it.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>