field: add update_reserved() function
Add update reserved function that replicates the behavior of the
update_bin() function.
This is in order to separate the reserved fields update user API syntax
from the other fields user API syntaxes, in particular: layout version
and eeprom id. The two above will use the same user API syntax as the
serial number (e.g. no delimiter between the bytes).
Therefore the change is for preparation of switching the update_bin()
function to use __update_bin() internal function instead of the
__update_bin_delim() internal function.

Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>