linux_api: change warning printing timing
In the function setup_interface(), we try to setup an interface
either through an i2c-dev file, or, if the first fails, through
an eeprom-dev file. Each failed attempt to access a file prints
out a warning message. If an i2c-dev file access fails but an
eeprom-dev file access succeeds, a warning message is printed
for the first, even though the program runs normally.
Print an error message only when both approaches fail.

Signed-off-by: Rinata Terno <rinata.terno@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>