command: do not write eeprom on error
Currently when the write command receives a non-existent field or an out of
range byte address, writing to eeprom occurs without the problematic field/
byte. This can cause a potential bug because it's not clear what exactly is
written. Change implementation so that no data is written to the eeprom unless
all fields/bytes are correct.

Signed-off-by: Rinata Terno <rinata.terno@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>