parser: limit acceptable i2c bus number and address
Both i2c bus number and i2c address can't surpass the size of the number 256,
yet any number is acceptable for these variables in the current parser.c code.
This can cause a bug when giving very large numbers that don't fit in the file
name array in the function setup_interface() in linux_api.
Limit the acceptable input for these variables to 256.

Signed-off-by: Rinata Terno <rinata.terno@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>