parser: split normal and error exit paths
When the user typed help/--help/-h or types nothing at all when running
the utility, the function cond_usage_exit() is called, the program exits
and a help message is printed out.
The same function is called when running the program with a syntax error.
In both cases the return value is 0 for success,even though the second case
is considered an error.
Split the two cases in to separate functions with appropriate return values.

Signed-off-by: Rinata Terno <rinata.terno@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>