parser: fix compilation for write disabled case
Commit f8166d8 (write range: add range writing ability) added is_bytes
argument to parse_new_data_stdin() and parse_new_data() functions.
When writing is disabled these functions are stubs.
The above commit only added the argument to the implementation of
write enabled functions implementation. This breaks the compilation for
write disabled case.
Fix by adding the arguments to the write disabled functions implementation.

Signed-off-by: Rinata Terno <rinata.terno@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>