command: change struct string_pairs to data_array
Currently the data to the command executer is passed in a 'struct
strings_pair' array. This design limits the possibility of using
different data structures for different commands. It is also forces
additional parsing of input to be made outside parser.c (e.g. the
function update_bytes() inside layout.c does a lot of parsing).

This change creates a new struct called data_array and uses it to pass
the data to the command struct. The new struct holds a size variable and
a union of pointers that can accommodate for a variety of data
structures types. This way the parser can create an adequate data
structure for each command.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>