parser: rewrite parsing data from stdin
The utility allows the user to input data from stdin for specific
commands. Currently, the whole process of reading and parsing the data
from stdin is done by the function parse_new_data_stdin(). The function
first saves all the input to a single buffer, then parses the buffer to
an array of strings, and then calls parse_new_data() to parse the array
of strings. Eventually, the function returns the parsed data as an array
of struct strings_pair, the same return data type of parse_new_data().

The problem with the current implementation is that stdin input parsing
is tied to a specific data type and a specific parser function. It
limits the possibility of using different data structures and different
parsing routines for different commands when input data is being passed
to the utility from stdin.

This commit takes a different approach on the subject. It separates the
process of reading the data from stdin, from the process of parsing the
data. A new function called read_lines_stdin() is responsible for
reading non empty lines from stdin and returning them in an array of
strings. Contrary to the current implementation, this one is indifferent
to the type of data coming from stdin. The new function does not call
any sequential parsing function, so it does not tie itself to a specific
data type or a specific parser function.

When reading data from stdin, the size of the input is unknown before
reading it to the end. We wanted to avoid setting fixed size limits on
the input, so each array or string is allocated to a fixed initial size
and can be reallocated later if needed. The initial size is defined as
the max size for conventional use of that allocated memory.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>