parser: new parser for 'write bytes' command
Currently, the input for 'write bytes' command is being parsed by the
function parse_new_data() to an array of struct strings_pair. The offset
(or range) is saved as a string in the 'key' field of the struct, and
the byte value is saved as a string in the 'value' field of the struct.
This strings_pair array later is passed to the command executer. Because
the array contains pairs of strings that are not yet fully parsed,
additional parsing of the input is needed before the data can be
verified and written to the EEPROM.

Moreover, when parsing the input data for 'write bytes' command, the
function parse_new_data() does some unnecessary data validation, like
checking if the offset number (or range numbers) is valid. The same
validation is done again later in the function update_bytes().
Regardless, this isn't suppose to be the job of the parser to validate
the data, only the syntax.

This change introduce a new struct called bytes_change and uses it to
pass the 'write bytes' input data to the command executer. The new
struct holds a 'start', 'end' and 'value' int variables. This change
adds a new function called parse_bytes_changes(). The new function is
used to parse the input data directly to an array of struct
bytes_change without first saving it in an array of string_pairs and
then parse it again.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>