parser: new parser for 'write fields' command
Currently the input for 'write fields' command is being parsed by the
function parse_new_data() to an array of struct strings_pair. The same
function was used before to parse input for 'write bytes' command, but
as of now it is used solely for 'write fields' command data input. This
function's signature and code are a relic of when it was used for
both 'write fields' and 'write bytes' commands. The function also uses
the not-so-safe strtok() function to split strings into tokens.

This commit replaces parse_new_data() with parse_field_changes(). The
new function specifically parses the input for 'write fields' command.

This commit renames struct strings_pair to struct field_change, and
renames the field 'key' to 'field' inside the struct. The new names
better reflects the data being saved to this struct.

This commit removes the functions valid_key() and alloc_cpy_str() from
the parser. Those functions are unused because they were referenced only
by the current parse_new_data() function.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>