linux_api: bug fix in i2c_read & i2c_write
To access bytes data on an i2C device, SMBUS offers a consecutive access
mode (byte after byte) and a nonconsecutive access mode (specifying the
address). Consecutive mode can be useful for EEPROM devices because they
auto-increment the address of an internal pointer after each access. The
drawback is that when more than one process is accessing the same device
at the same time, they both increment the same internal pointer with
each access.

The 'size' parameter of i2c_smbus_access() sets the access mode:
size=I2C_SMBUS_BYTE to access consecutive bytes.
size=I2C_SMBUS_BYTE_DATA to access a byte on a specific address.
The 'command' parameter holds the specific address on the device.
In the consecutive mode, the function ignores the value stored in
'command' because it uses the internal pointer of the device.

The function i2c_read() uses the consecutive access mode although this
probably wasn't the original intention. The function sets the 'command'
parameter with an incrementing value after each read, but because the
access mode is consecutive, this value gets ignored and the function
reads from where the internal pointer of the EEPROM is pointing.

Normally the utility is able to read a complete EEPROM consecutively
without any problems, but in some rare cases, like when a background
process is accessing the same i2c device in a way that modifies its
internal pointer, it can interfere with the read process of the utility
causing it to read incomplete data.

The solution is to simply change the value of 'size' in i2c_read() from
I2C_SMBUS_BYTE to I2C_SMBUS_BYTE_DATA. In addition, now there's no need
to reset the internal pointer before the reading loop. This reset can
potentially interfere with other background process, so it was removed.

The function i2c_write() does not use the consecutive access mode, so
the same fix is not needed there. However, the function does reset the
internal pointer of the device before the write loop. This reset is
unnecessary and was removed.

Sources: i2cdump man page and source code
	 www.i2c-bus.org/auto-increment

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>