field: improve value processing in update_date()
Use strtoi() to process value strings instead of strtok() and strtol().
Validate the syntax better. Print a proper error for invalid month.

strtok() modifies the original value string in a way that caused the
function to accept values with wrong syntax (like 1////Jan/18 or
1/Jan/18////). It also prevented printing the original input value in
case of an error because it got modified.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Acked-by: Nikita Kiryanov <nikita@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>