field: improve value processing in update_bin_ver()
Use strtoi() to process value strings instead of strtok() and strtol().
Validate the syntax better and print a proper error for invalid value.

The function is said to "strictly enforces the data syntax" in the
description, but in practice it accepted invalid values causing
unexpected behaviors. Examples of values the function accepted:
1.2222 (minor version is limited to two digit) - saved 23.22 to field
-1.999 (version must be positive) - saved 8.99 to field
1.-99 (version must be positive) - saved 0.01 to field

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Acked-by: Nikita Kiryanov <nikita@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>