common: pass base param to strtol() in strtoi()
Add a base parameter to strtoi() and pass it to strtol().
Rename strtoi() to strtoi_base().
Add a new strtoi() function that calls strtoi_base() with base=0.

strtoi() is a wrapper function for strtol(). The function strtol() has a
parameter to determine the numerical base used for the conversion. If
the parameter is 0, the base gets determined by the format. Passing this
parameter to strtoi() isn't necessary for current uses, but it will
allow this function to replace safe_strtoui(), a similar function used
by a few functions in field.c.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Acked-by: Nikita Kiryanov <nikita@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>