parser: correct the accepted address range
Change the range of an accepted i2c address parameter to be 0x03 to 0x77
instead of 0 to 255.

The address of an I2C transfer is 7 bits long, so it theoretically
allows 128 I2C addresses. However, some addresses are reserved for
special purposes. Thus, only 112 addresses are available. The available
addresses are in the 0x03 to 0x77 range.

sources:
  www.i2c-bus.org/addressing
  i2c-tools source code (function parse_i2c_address() in i2cbusses.c)

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>