parser: change input parsing functions signature
Change the signature of the input parsing functions to accept a pointer
to a data array and to return an int value indicating success or error.

The affected functions:
parse_field_changes()
parse_bytes_changes()
parse_bytes_list()

This change is needed for an upcoming feature where the size of the
data array will change during parsing. By passing a pointer to the data
array, the function can change its size value.

This change also improve error propagation and code readability.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>