linux_api: fix checking if EEPROM device exists
Use the 'errno' value after the call to stat() to check if the device
exists, instead of using only the return value of stat().

stat() can fail (return value -1) for many reasons, not only because the
path doesn't exist (permissions, memory error, etc...), so checking the
'errno' value is important.

Add printing of a "failed accessing" error if stat() failed for reasons
other than a non existing path, similar to the error message when
listing I2C devices.

Return success only if device exists and accessible.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>