layout: fix 'fields' commands on an unknown layout
Issuing a 'write fields' or a 'clear fields' command when the layout is
unknown results in a segmentation fault.

Execution example:
$ ./eeprom-util write fields -l raw 6 0x20 potayto=potahto
Segmentation fault (core dumped)

This happens because the array "layout_unknown" has no valid fields and
thus the short name of the single "raw" field is NULL, so the call to
strcmp() in find_field() aborts with a segmentation fault..

To solve this, add a checkup in find_field() to test if layout->fields
is "layout_unknown". If so, print an error and return NULL.

In relation to this, because the "raw" field of "layout_unknown"
shouldn't be writeable, set the update function to NULL.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>