field: encapsulate the function pointers of field
The current implementation relies heavily on mutual interaction between
the field object and the layout object. This is due to the bad habit of
allowing the layout object to initialize and access the private members
of the field object. We want to change this behavior by encapsulating
the use of field object.

This patch addresses the initialization of field's function pointers by
the layout object for all the predefined fields arrays.

To solve this, add the ability to initialize the function pointers of
a field by selecting the field's type and then calling a field init
function. The function will set the function pointers of a field
according to the field's type. All the available field's type will be
listed on the field's header file.

This change will spare the need for setting the function pointers of
each field for each of the predefined layouts in layout.c. Instead, the
predefined layouts will set only the field's type, thus improving the
encapsulation of field object.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>