field: make all ops functions private
The ops function of the field object are public because they were used
by the layout object to initialize the fields of the predefined layouts.

Since the previous commit, the initialization of function pointers has
moved to field.c from layout.c, so all the ops functions are no longer
have to be public.

Set all ops function to private and remove the functions declaration
from the field header file.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>