field: encapsulate data initialization of field
The current implementation relies heavily on mutual interaction between
the field object and the layout object. We want to change this behavior
by encapsulating the use of field object.

This patch addresses the data initialization of each field by the layout
object. Currently it is done by accessing the private members "data" and
"data_size" of the field object.

Move the data initialization to the field init function and add a getter
function for the size of the data. This replace the accessing of private
members thus improving the encapsulation of field object.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>