layout: replace NULL names with string names
Following the recent changes, field's names are no longer required to be
NULL in order to identify the field as a reserved or a "raw" field.

Change all the NULL names to string names in the predefined layouts.

This change will save a bit of code in a following commit.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>