field: encapsulate accessing field names and type
The current implementation relies heavily on mutual interaction between
the field object and the layout object. We want to change this behavior
by encapsulating the use of field object.

This patch addresses the access of fields names and type by the layout
object while searching for a field in the layout.

Add a function to field object that checks if any of the field's names
matches the given name and use this function in the layout object. This
change improve the encapsulation of field object.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>