field: separate print field value and field name
Change all the functions that print the field's name and value to print
only this field's value. Add a function that will print the field's name
and then will call the function that prints the field's value.

This change improves code reuse and makes it easy to support more print
formats in an upcoming feature.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>