api: separate initialization from interface setup
Currently the function setup_interface() does more than just setup an
interface to an I2C device, it also initializes the api object. Some
commands don't need to setup an interface, so calling setup_interface()
before executing the command is counter intuitive. We want to setup the
interface only if and when it's needed.

Change setup_interface() to only setup the interface, make it private to
the api object and set it to be called only when needed.

Add api_init() to initialize the api object. This function can replace
every existing call to setup_interface() even though it doesn't setup
an interface, because of the changes described above.

The benefits of this change:
1. Interface related errors will occur only when accessing a device, and
   not also when the object initialize, thus simplifying error handling.
2. No need to provide setup_interface() with invalid (negative) I2C bus
   or address to setup the probe method, so the function can ASSERT the
   validity of the I2C bus and address, and it also sparing the need to
   set default invalid values by the parser.
3. Provide a more object-oriented implementation.

Signed-off-by: Yaniv Levinsky <yaniv.levinsky@compulab.co.il>
Signed-off-by: Igor Grinberg <grinberg@compulab.co.il>